Delay restoring state until we have an adapter
Delay restoring of mCurrentItem and adapter state in
onRestoreInstanceState until there is an adapter. When setCurrentItem is
called before the adapter is set and after onRestoreInstanceState has
run, the pending value is updated to reflect the new current item when
possible.

If the restored value is larger than the adapter size, it is clamped to
a valid value, similar to what happens in setCurrentItem.

Bug: 127838185
Test: ./gradlew viewpager2:connectedCheck
Change-Id: I92c80abfd51fff337e1d892e693dd9b83b4aa1f8