Replace boolean handleOnBackPressed with enabled state
Instead of having handleOnBackPressed() return
a boolean stating after the fact that it handled
the back button, add an enabled state to each
OnBackPressedCallback. This allows callbacks to
set whether they support back button events
ahead of time.

As a consequence of this, OnBackPressedCallback is
now a class rather than an interface.

Test: Updated tests
BUG: 129946737
Change-Id: Ibcdce43b408cf2898735a977825b9efe925e4db9