Remove TextStyle.inherit
This attributed was ported before. But we can't support the expected
behavior when TextStyle.inherit = false.
For example,
text = TextSpan(
    text = "hello",
    style = TextStyle(fontSize = 8),
    children = listOf(
        TextSpan(
            text = "world"
            style = TextStyle(inherit = false)
        ),
        TextSpan(text = "!")
    )
)

In the above case, inherit = false specifies that the 'fontSize = 8'
set on top should be unset for "world".

This behavior is not supported by SpannableString. So the only way
we can make it work is to set a span to cancel the effect of
'fontSize = 8'.

However, developers can always achieve the same result by breaking
the top level span into two pieces. So this feature is removed and
won't be supported.

Bug: 121157451
Test: ./gradlew test

Change-Id: I23d75cacba5290341b085534f7dcbb649628caa3