Call setMaxLifecycle() after add() in FragmentStatePagerAdapter
SetMaxLifecycle() must be called after add() or an
IllegalArgumentException will be called because the fragment is not yet
attached to the FragmentManager.

In instantiateItem(), if used, the call the setUserVisibleHint() will still
happen before the fragment is added. If only the current fragment is resumed,
setMaxLifecycle() will be called after add().

Test: support v4 app
BUG: 129780800
Change-Id: Ic3406fcb83554ba94e270f8cffd8acacf3a2da4f