Account for delays when sending drags in FakeDragTest
In FakeDragTest, calls to ViewPager2's fake drag methods are scheduled
such that it results in a smooth drag. But the actual time at which
those callbacks are executed is unreliable (too late), which influences
the velocity calculated by Viewpager2. Sometimes this results in a
velocity above the fling threshold when a velocity below the threshold
was expected, which causes a fling to happen when it shouldn't.

Fix this by posting callback i after executing callback i-1, so there
will be a guaranteed delay between the two. Also, actively suppress
flings when we don't want them by extending the fake drag with 100ms of
zero motion. That guarantees the VelocityTracker returns a velocity of
0.

Bug: 129296837
Test: ./gradlew viewpager2:connectedCheck
Change-Id: Ib0a8f057a80ea20280029899545691e4c78e2669