Only set mDragStartPosition when not yet set
When a drag is started in ScrollEventAdapter, the position where the
drag started should only be set if it wasn't set before. If it is was
already set, the new drag is taking over from another (fake) drag and
the start position shouldn't change.

Added test case that teases out the bug.

Bug: 131411265
Test: ./gradlew viewpager2:cC
Change-Id: Ied2c508e97ffeb093570b3ffe079485d278e56f7