Avoid child Fragment animations running during parent Fragment transactions
When a child fragment is on top of the parent and that parent is replaced,
the child fragment's animation from the most previous operation is
executed. This can cause unexpected behavior in developer animations.

Just clearing the child animation results in the parent fragment
appearing on top of the child as the animation is executed. This is
because the child fragment's view is removed before the parent animation
is executed.

This change ensures that if a parent fragment is being removed, its
child fragment's view remains, and is destroyed with its parent's view.

Test: added test, tested visually in app
BUG: 121017790
Change-Id: I4f008316ea00258f9bec445216f1b612330d509a