Change setMaxLifecycle to only be one Op
SetMaxLifecycle consists of two operations (a set and unset). There is
an issue with adding a fragment and setting the max lifecycle where if
allow reordering is not set to true, a Fragment goes to the RESUMED
state before being capped properly. This happens because the unset Op is
executed first, although it is unnecessary for a new Fragment.

Instead of using two Ops, this change allows a single Op to maintain the
new max lifecycle as well as the previous one. When setMaxLifecycle is
called, the Op's currentMaxLifecycleState is set to the given state, and the
oldMaxLifecycleState is set to the given Fragment's maxLifecycleState.
When there is a pop from the backstack, the Op's oldMaxLifecycle is set
as the currentMaxLifecycle.

Test: edited current test, reran fragment test
BUG: 131630931
Change-Id: Id94f991ea2710e64ff51a1f130ee680cab9541d2