Ensure destinations are linked properly on nav back stack
Currently when we navigate to a destination, if its ancestors are
already on the nav back stack, we do not re-add them. This can cause
an issue when we save and restore where destinations on the back stack
are not reachable, causing an unknown destination IllegalStateException.

When navigating to a destination, we should add the ancestors needed to
get to that destination from the current one to the back stack. This
ensures that on restore we have the proper path to access all
destinations and rebuild the entire navigation graph.

Test: test app, new test, reran nav test
BUG: 131733658
Change-Id: I6f07fb3ba2ccf2c691b66df304a4f3c160313d82