Fix invalidation handling
Composing an recompose scope incorrectly updated the invalidations
This could lead to pending invalidation being skipped. In b/132638679
this lead to switch components not being updated to their correct
visual state as transitions were skipped.

Invalidations that survive a composition might have the wrong offset
in the next invalidation causing an "Expected a group start" error.
In b/132638679 this lead to an exception.

Fixes: b/132638679
Test: ./gradlew test connectedCheck
Change-Id: If09087a52b566dd3ed0dc6d3ba01f9822f92d37c