Update DrawerLayout gesture navigation behavior
The main change is that the drag edge size now matches
the system gesture area widths.

This required changes to both ViewDragHelper and
WindowInsetsCompat, which are currently restricted.
They will be un-restricted soon.

Removed the DrawerLayout test package, since this
behavior is extremely hard to test since it relies on
system settings.

Test: manual
BUG: 128547310

Change-Id: I3f2694decc56bd1416cffe4fd6e3057037d57c28