Remove androidx.collection from room-runtime.
This requires also removing androidx.core since that would bring
collections as a transitive dependency.

* ArraySet becomes HashSet
* ArrayMap becomes HashMap<K, V>
* SparseArrayCompat becomes HashMap<Integer, V>
* ActivityManagerCompat.isLowRamDevice() is copy & pasted to a private
  static util method.

Special consideration had to be taken for migrations since when finding
a migration path Room prefers migrations that gets the database to the
target version faster (i.e. migrations that jump), thus needing a
sorted key map, such as a TreeMap.

Bug: 134699078
Test: Existing test pass
Change-Id: I67527d68ae9676fd21c7c016af10d0b2ca1289e1