Only keep baseline if the fine has more than a header.
Previously, we would try to store baseline files even if they only
had a header and no actual issues. Since that is not helpful, this
CL changes to check if there is more than header.

Disable HiddenSuperclass check as androidx does not have an issue with
hidden parent classes.

Also rerun the task to make sure all the ignore files are up to date.

Test: ./gradlew ignoreApiChanges
Change-Id: I8c20ce8c9dabd6797b42b97c63a08cce938e5d72