Created factory method for BlendMode Color Filter
Separated PaintCompat helper methods to enable creation of
a color filter with a provided color and BlendModeCompat parameter.
This factory will create either a BlendModeColorFilter or a
PorterDuffColorFilter based on capabilities of the given API level.
A BlendModeColorFilterCompat API is not feasible as the underlying
framework primitives are package private, so a static factory is
created instead.

Moved mapping logic out of PaintCompat into BlendModeUtils to be shared
between ColorFilterCompat and PaintCompat

Test: Updated PaintTest to leverage underlying ColorFilterCompat and
BlendModeUtils APIs
Bug: 135943149

Change-Id: Ifd99f8b406744e4e165d268d75fe650e89917e0b