Flatten text input composables
Flatten text input composable for following two reasons:
- Custom text input field idea was pushed back
TextInputClient composable and Focusable composable were introduced for
supporting custom text input field for directly communicating with
InputMEthodService. However, this project was pushed back and no longer
used or targetted for public APIs.

- InputEvent propagation workaround
Looks like if two or more PressGestureDetector are nested, only one of
them is called. To avoid this issue, flatten the Focusable and
InputFiled composable and use the single PressGestureDetector as for
now.

Bug: 135556976
Test: ./gradlew test
TEst: Manually done w/ observing logcat
Change-Id: I624228cd929b85b17c274c4399c656ed54f95e0e