[ImageCapture] Don't wait 3A convergence when AF mode is OFF Or 3A state is UNKNWON
Don't wait 3A convergence when AF mode is OFF Or 3A state is UNKNWON
In most devices, front camera has no auto-focus capability. In this case,
AF mode is off and we should wait for AF to be locked or focused.

Some devices don't return AE/AWB state in CaptureResult, so we should ignore the check
if AE/AWB/AF states are unknown.

Bug: 135963434
Test: All CameraX test
Change-Id: I8275307c4295bce130c31fc45eb0fdb8d608afb7