Extract biometric prompt demo logic to controller
Currently, all of the logic for initializing and launching the prompt
for BiometricPromptDemo lives inside the demo activity. This commit
extracts the majority of this logic to a new BiometricPromptDemoController
class. This will allow it to be shared with the fragment-based demo
implementation that will be introduced in a later commit.

Test: Launch the biometric prompt demo app, try various combinations of
options, and verify everything works as before.

Bug: 131980596
Change-Id: I0e1f9b321c1e9b81e05ed79aa76f0a34e68b7948