Catch IllegalArgumentExceptions in NetworkStateTracker#startTracking.
This seems to happen with Nvidia SHIELD K1 Tablets.  It's similar to
our workaround for #stopTracking.  Since this is a device-specific
problem, the best we can do is not crash in the background and let
JobScheduler's implementation execute this worker.  (This appears to
occur on API 24 devices, which means JobSCheduler can be used.)

Fixes: 136569342
Test: N/A
Change-Id: I1e1351ee41a45cdd67a6b73846ffc129edd29004