Introduce AnimatedDraggable.
This CL makes possible to glue together AnimationFloat and drag gestures
easier by creating AnimatedDraggable abstraction.

In a lot of components, such as Scrollers, Drawer, BottonSheet, SeekBar
you need to be able to drag something and also animate it when drag ends
or button pressed. Also, sometimes you want to drag only between
predefined set of points and don't want to stay between them. There's,
to simplify, 3 usecases for animations with drag support:

1. To be ablt to fling after drag has ended with velocity.
2. To be able to "emulate" drag with button press (e.g. Drawer open with
button or "page down" in scroller)
3. To be able to end drag only in predefined set of points AKA anchors
(Drawer with 2 states or SeekBar with N predefined volume settings)

It's hard to create component for each of this, as you usually need 2
of this 3 or even all 3 features. This CL adds AnimatedDraggable that
makes it possible in controlled and declarative fashion.

Bug: 137012917
Test: To be written
Change-Id: Ic699d217b7fb193005885448e35ff925946079c7