Ensure that animations finish before dispatching destroying the view
If a fragment whose view is being destroyed has a valid animation we
should wait until that animation is completed before destroying the
view.

This change checks if the exiting Fragment has an animation associated
with it. If it does not, destroyView is dispatched, otherwise,
destroyView is skipped until the onAnimationEnd callback is reached.

Test: Changed AnimationTest, all test pass
BUG: 136110528
Change-Id: Id598c955712056189cc1ef76487f5d99fd0b3430