Ensure Fragments that have views call onDestroyView before onDestroy
If a Fragment has created a view it should ALWAYS call onDestroyView
before calling onDestroy.

This change adds a check in StrictViewFragment to ensure any createdView
is destroyed by the time onDestroy is called. If animations are ended
prematurely by a call to saveAllState(), the view is destroyed right
after the animation is canceled.

Test: Changed Test, all tests pass
BUG: 136110528
Change-Id: I3b536a412666f4d185155e9e2185f5f2e267c22d