Add attributes that allow marking types as stable
Added `@Immutable` and `@StableMarker` and marked
`@Model` as stable.

A stable type can be compared during composition for
equality. An unstable type is always assumed to be
different and is never compared or added to the slot
table implicitly. Whenever any of the parameters of
a component is unstable then none of the parameters
are compared or added to the slot table.

Removed the runtime checks for valid comparisons as
the type defintion now controls which values are
compared and non-stable values will not generate
a changed or update

Test: new tests, ./gradlew test

Change-Id: I5003c98db1853c0875480521a67e6cf18f50eec1