Call onTrackInfoChanged when current media item changes
It makes MediaPlayer call onTrackInfoChanged when the current media
item chages as well. It gives developers a clear guide about the
lifetime of List<TrackInfo> so that they could call selectTrack with
valid TrackInfos.

Also, it enables MediaController to reset their cached stale tracks.
Without this, MediaController should be able to distinguish MediaItem
instances at onCurrentMediaItemChanged as the callback is called for
both when
1) new MediaItem became the current item, and
2) MediaMetadata is updated for the current item.

Bug: 138081547
Test: ./gradlew :media2:media2-player:connectedCheck \
                :media2:media2-widget:connectedCheck
      (on API 26, 28 emulators and API 29 device) &&
      ./gradlew :media2:media2-player:check tipOfTreeDocs
Change-Id: I46922d820657f82d04b7b00761f5cd11c04ce219