Fix mCurrentItem and events when data set changes
If the data set changes during a smooth scroll and the first visible
item is removed in this change, the smooth scroll may land on another
page then expected.

Fix this by dispatching a selected event and an idle event when RV
transitions to idle and we are not yet idle (and a few more specific
conditions).

Bug: 137642608
Test: ./gradlew viewpager2:cC
Change-Id: Ia241d00aa7758ddbdd937a97aa17a15ed8feb343