Add more tests for data changes while scrolling
Creates a new parameterized test to tease out different modifications of
the adapter's data set while smooth scrolling. Note that while the
expectedFinalPage and expectedFinalPageText are not what you'd expect
intuitively, it is fine to hardcode them as they are not determined by
ViewPager2, but by an external and constant factor (RecyclerView).

Bug: 137642608
Test: ./gradle viewpager2:cC
Change-Id: If08869bcb45c014168f9f208457894d227ab7588