Material Button Styles
Introduces a Button component that is configured via a ButtonStyle. ButtonStyle is a data class that holds on to the configurable parameters: color, shape, border, elevation, paddings, default text styles. We then provide 3 implementations of ButtonStyle: ContainedButtonStyle, OutlinedButtonStyle and TextButtonStyle. When a developer wants to style a button differently, they can create their own ButtonStyle and apply it to their buttons.

Bug: 113845528
Test: few more tests for Buttons styling
Change-Id: If41c79898658ee124dfb445c503bab935647a881