Fix saveWhileAnimatingAway() local test failures
FragmentAnimatorTest saveWhileAnimatingAway() test can fail locally
because of animations ending before asserts can be reached.

We should not wait 2 cycles after executing pending transactions on the
FragmentManager. Executing pending transactions forces the animation to
begin and by waiting 2 cycles, it is given a chance to end. Instead,
after we executing pending transactions, we should ensure animations
have been started, then do our asserts. This ensures that even if the
animations were to run with a duration of 0, we still would do our
asserts before the fragment gets marked as inactive. So we should no
longer have to rely on timing for this test to be successful.

Test: all tests pass
FIXES: 138798649
Change-Id: I43f44c535f02ca8af19dce5c787542a7a22f6284