Introduce max lifecycle to NavBackStackEntry
Instead of all NavBackStackEntry instances
being the same state (equal to the NavHost's
state), allow them to be updated as the user
navigates between destinations.

This comes in two forms:
- When you navigate from one destination to
a FloatingWindow destination, the previous destination
is paused (i.e., moved to STARTED)
- When you navigate from one destination to
a not FloatingWindow destination, the previous
destination is stopped (i.e., moved to CREATED)

This ensures that the state of the navigation
graph NavBackStackEntry instances is the max of its
children and that those children are updated
correctly.

When updating the state, we make sure that changes
to the navigation graph and its child destination
are properly nested: graphs are resumed before
the child destination is resumed and the child
destination is paused before the graph is paused.

This is done specifically in
dispatchOnDestinationChanged() so that the state is
not erroneously changed earlier due to
popBackStackInternal calls, avoiding cases where
we stop NavBackStackEntry instances unnecessarily.

Test: new NavBackStackEntryLifecycleTest
BUG: 36818088
Change-Id: Ie76db4cf8ace5922dd8e8591d727c2e909bf3f70