Make NavBackStackEntry and getBackStackEntry public
Internally, when using getViewModelStoreOwner,
the hidden NavBackStackEntry would hide the fact
that it implements SavedStateRegistryOwner or that
it implemented HasDefaultViewModelProviderFactory,
making it difficult for developers to determine
what capabilities they'd get out of the returned
ViewModelStoreOwner.

The current API also meant that in order for
developers to use a non-default Factory that extends
AbstractSavedStateViewModelFactory, they
would need to inspect the source code and unsafe
cast the returned ViewModelStoreOwner to a
SavedStateRegistryOwner to ensure the correct
saved state ownership. They'd also be unable to access
the arguments originally passed to the navigation
graph.

By making getBackStackEntry() public and the underlying
NavBackStackEntry class as well, developers will
be able to have the same level of access to the
underlying SavedStateRegistryOwner as we have.

In addition, this also makes the LifecycleOwner of
navigation graphs available, opening up the ability to
observe the Lifecycle at the navigation graph level.

Test: ./gradlew checkApi
BUG: 36818088
Change-Id: I725744a03875f0ea1df30aba6c6df74ee7f799d9