Initial commit of PagePresenter
Test: ./gradlew paging:p-c:test

PagePresenter recreates one of the core roles of PagedList in
isolation - holding data, and dispatching updates.

It operates on PageEvents, a simplified format for data loaded from a
paging system to be presented.

The long term goal of this code is to be the sole recipient of a
stream of events from another system that manages fetchs from a
PagedSource, so that the stream can be manipulated to alter
presentation of data to the adapter part of paging. It still presents
a list with list update callbacks, but that's all it does.

Change-Id: I7b453bd255705646c27591a916a9b4cf22a3361d