Allow NavHostFragment to be added without a container
NavHostFragment is designed solely to host child
fragments that contain a View. In almost all cases,
this means the Fragment must be added with a
container ID.

However, if you add a NavHostFragment using
ViewPager2, it does *not* add the child fragments
with a container. In this case, NavHostFragment
should use a predetermined R.id for the container
hosting its child fragments.

Test: new test passes, sample project works now
BUG: 139421197
Change-Id: I24be55c2edaebf3e0747737c66e7e918f1dc3108