Change FragmentContainerView draw order when popping
Currently FragmentContainerView always draws disappearing views first so
they appear below any views that are entering. When views are being
popped they are exiting, but the expected behavior is that they exit on
top of the view that is returning as a result of the pop.

This change introduces a method in FragmentManager that allows the
default FragmentContainerView drawing order to be reversed so that
exiting views are once again on top. This method is called when
Fragment are removed from the FragmentManager. If the removal is not
caused by a pop, the drawing order remains the same, if the removal is
caused by a pop, the drawing order is reversed.

Test: Added tests, tested in app
Bug: 139104187
Change-Id: I6cbbc91ad84b21197dcfc2776d602070362f44f4