Fix broken test ProcessingSurfaceTextureTest.writingToSurface()
ProcessingSurfaceTextureTest.writingToSurface is broken on postsubmit. The issue is that, the test blocks on handler being idle and then verifies the received frame. But there is a race condition that we could be trying to block even before the handler becomes busy. The fix is using a semaphore instead of relying on thread being busy.

This might not be the cause of the postsubmit failure, but it's at least one step in the right direciton.

Bug: 139653545
Test: manual test and ./gradlew bOS
Change-Id: I86e49153f5bf518fa42ebce83117db6621c45eab