Don't ignore requestDisallowInterceptTouchEvent in SRL
A SwipeRefreshLayout ignored calls to requestDisallowInterceptTouchEvent
when nested scrolling was not enabled in its child, but there is no good
reason why it shouldn't propagate the call to its parents. It is OK for
SRL itself to dishonor the request, as one can always use
OnChildScrollUpCallback to control cooperation with scrolls in children.

Bug: 138314213
Test: ./gradlew swiperefreshlayout:cC
Change-Id: I4959df942ebb2097f3c3ebe39ff0582f6ae6f395