Fix incorrect pop animation in multiple transactions
When a fragment is added by a pop and then removed in back to back
transactions, we should use a normal exit animation instead of a popExit
animation.

This change keeps track of the first Fragment being visibly removed
(remove, hide, or detach) in a set of FragmentTransactions. It passes
the animations of subsequent removing fragments to the one that will be
visible when the animation is executed. The Fragments are tracked by a
tag on their container.

Test: all tests pass, and tested in app
Bug: 111659726
Change-Id: I18eefc2fdb4fc4c27c467212cc2e68beb20ea00b