Make "rejecting-state" case working in TextField
It is allowed not to set state TextField in the app but it becomes
inconsistent state between the source-of-truth and IME or TextField
internal state. To synchronize the internal state with application's
state, introduce new variable that maintains the generation of the
onValueChange callback. By this generation state, every time the
TextField calls onValueChange callback, compose happens, then
synchronize the internal state with app's state.

Bug: 139821920
Test: ./gradlew :ui:ui-framework:connectedAndroidTest
Change-Id: I513e61aa1a67c40b5b8d7eee171e6bf1f00dfeec