Remove timestamp from CarUxRestrictions.
This is a shim class that interfaces the framework class, removing it
since it can no longer access the API on Q+.

getTimestamp() returns the time at which this UX retriction event was
deduced. As part of Auto API review, the framework API getTimestamp()
was changed to @hide in b/126757814 as it was deemed unnecessary for
consumer.

Bug: 140267580
Test: ./gradlew :car:connectedCheck

Change-Id: I73523b81e58285c311355704c6e1e3ade70f8b1e