Fix test timeout issues for RecyclerView.
Timeout issues seem to be introduced by aosp/884635
which made a bunch of tests that were @MediumTest,
@SmallTest, which may have been a bit too aggressive
for older (and slower) devices.

Most of the timeouts seen to be be occuring in
ItemAnimatorV2ApiTest, so this CL just addresses those.
There seem to be more flaky failures in
ScrollToPositionWithAutoMeasure and
WrapContentBasicTest, but will leave those alone for now.

Fixes: b/139653729
Test: Just changed @SmallTest to @MediumTest, should pass.

Change-Id: I750f502b52f651f695621734c19fbc5a294dccca