Add documentation around setCustomAnimations
In a FragmentTransaction, setting custom animations is only applied to
operations after setCustomAnimations is called, not the entire
transaction or any of the operations prior to the call to
setCustomAnimations. This means that operations in the same Transaction
are allowed to have different custom animations.

This change adds documentation around how setCustomAnimations actually
works.

Test: ./gradlew generateDocs -PofflineDocs=true
BUG: 141682906
Change-Id: Ic7a35093322a04b36f0cc6ce9b3c5d49c2e7429f