[webview] Correctly support null WebViewRenderProcessClient.
If the render process client is set to null (the expected means of
resetting the client) then this could lead to a NPE when a callback was
invoked from the support library glue. We should avoid creating an
adapter object (and pass a null InvocationHandler, which is what the
glue code expects) instead.

This CL also adds a test that setting a null render process client performs
as expected.

Bug: 141776033
Test: ./gradlew webkit:connectedCheck \
Test:   -Pandroid.testInstrumentationRunnerArguments.class=\
Test:   androidx.webkit.WebViewRenderProcessClientTest\#testSetNullWebViewRenderProcessClient
Change-Id: I5ea0daa6796bf54c13c2242e0710821b264f2dd7