Move create/destroy to FragmentStateManager
Due to restoreSaveState()'s cleaning up
of no longer valid retained Fragments, moveToState()
now always creates a FragmentStateManager if one
does not exist so that we can continue to run the
create callbacks despite these retained Fragments
not being in mActive.

Interactions with mExitAnimationCancellationSignals
as part of destroying the Fragment remain in
moveToState() for now.

Test: Fragment tests still pass
BUG: 139536619
Change-Id: Id05f4e78f6fc0cef6be9cc7abc0d58b40b2e3f43