BUGFIX: crash on exiting the test app.
The test app crash for about 1 in 20 times when home button is pressed.

Originally introduced in aosp/980247, the root cause is a framework bug where abortCaptures() on legacy device throws NPE. aosp/1105424 was an unsuccessful attempt to work around the issue. Since the abortCaptures() change is an optimization to begin with, we will simply disable it on legacy devices. Also reverted the unneccessary workaround aosp/1105424.

Bug: 139448807
Test: manual test and ./gradlew bOS
Change-Id: I52a91ac0e6b3ed7b183547c77906f942bd37799c