[ImageCapture] Calls onCaptureCancelled when camera is closing, also cancels ImageCaptureRequest all when onStateOffline.
This is to fix the issue that ImageCapture takePicture request is blocked if takePicture() multiple times and
pressing home. There are two kind of issues and its solution shown as below:
(1) Capture request is onCaptureCompleted, but cannot get the Image from ImageReader:
   * Adding onStateOnline / onStateOffline callback to UseCase, Camera will notify these callback when UseCase
   becomes online or offline. These callback will run in mainthread.

   * In ImageCapture, abort all image capture request when onStateOffline, and calls onError to notify apps.

(2) Ensures pending CaptureConfig's onCaptureCancelled is called when Camera is closing.
   * When closing camera, cancels all issued capture request(but not in camera2 pipeline yet) in current
     CaptureSession.

Test: all tests
Bug: 135582703
Change-Id: I21bac7e590672fe268aac54cbe9b17e369333240