Improve how we track models used during the node's drawing/measuring/layouting
Previously we had some incorrect assumptions like we were considering measuring and layouting as one step, so we would clear the old map of models to nodes and then start measuring assuming the layouting is always happening after. It is true, but sometimes layouting(positioning) can happen without measuring so we should actually track models used during the positioning completely independent from the nodes used during the measuring. In the bug reproduced in Drawers we had related bug: drawing and layouting were using the same model, and first layouting happened - we remember that we use the model during this stage and then before starting drawing we were clearing the map for this model so the next layouting wasn't happening as we already forgot we used this model. The only way to solve this in a predictable way is to completely separate maps for this three stages. So I introduced three separate NodeModelMap - my small abstraction combining both modelToNodes and nodeToModels ObserverMaps.
I covered it with quite a lot of tests, not all of them were failing before the change but they would help us to enforce the correctness of the model changes tracking. I also covered the code with few more `require(layoutNode.isAttached())` assertions. So we first wouldn't be able to ever request measure/layout/draw for the detached node and then to really perform this. Previously it was possible to call measure or place on a dettached node(when owner is null) which seems wrong and could be masking new issues(for example if measuring happened without an owner this means models used during the measure wasn't saved so this node wouldn't be remeasured when the model changes).

Test: Drawer works with models used in layout block. New tests
Bug: 142632149
Change-Id: Idc8a9be43fd40e852abb7ab562c1f443dc94388c