Change canSetSupportedResolutionsToConfigTest size as MediumTest
When the tests are running in presubmit, it is the only test in the
file. Therefore, it will take some extra time to load mock library and
then may exceed 300 ms timeout threshold on Pixel API 28 device. When
running locally with other medium test, it may only take less than 100
ms. Therefore, enlarge it as MediumTest since it is caused by the test
environment.

Bug: b/141965720
Test: ./gradlew camera:camera-extensions:build && ./gradlew camera:camera-extensions:connectedAndroidTest
Change-Id: I6e1574a79e04a450ce10daa5dba7decddd36a20c