Fix for drawing sometimes happening without observing the model reads
The issue were happening because in methods like onEndLayout and onEndMeasure we had the code currentNode = layoutNode.parentLayoutNode. But it is not always correct. For example if we just call layout on node1 after the measuring we would have currentNode = node1Parent, but it is should be null as we ended all the stages. Then there in RepaintBoundaryView.dispatchDraw we were starting model changes observing only when ownerView.currentNode == null, but as we discovered easrlier sometimes layout or measure stages could not clear currentNode value so the drawing were happening without the model observing and transition were not causing invalidation as a result.
In general the code about model changes observing became hard to maintain and was hiding a lot of small mistakes which are hard to reproduce. So I decided to refactor this and extract the logic into separate NodeStagesModelObserver, add a lot of assertions to enforce the contact and never experience issues like this and covered this class with tests for the cases which hard to cover as an integration test. And I already added more integration tests for this in ModelReadsTest.

Bug: 142852849
Test: animation in the Rally app is fixed. new NodeStagesModelObserverTest
Change-Id: Ia1fa0be9f36789dc54a1207dcb09474b6d1feab6